requirements.txt
tqdm
pandas
numpy
scipy
hydra-core
omegaconf
networkx
scikit-learn
pytorch==1.5
torch-geometric==1.6.1
torch-sparse==latest+cu102 -f https://pytorch-geometric.com/whl/torch-1.5.0.html
torch-scatter==latest+cu102 -f https://pytorch-geometric.com/whl/torch-1.5.0.html
torch-cluster==latest+cu102 -f https://pytorch-geometric.com/whl/torch-1.5.0.html
torch-spline-conv==latest+cu102 -f https://pytorch-geometric.com/whl/torch-1.5.0.html
